Fix: Store refs as 16 bits, not 8
Refs are handled as 16 bits properly everywhere, yet are stored as 8
bits in the assembly.

+++ src/smsutil.h
@@ -353,7 +353,7 @@ struct sms_assembly_node {
 	struct sms_address addr;
 	time_t ts;
 	GSList *fragment_list;
-	guint8 ref;
+	guint16 ref;
 	guint8 max_fragments;
 	guint8 num_fragments;
 	unsigned int bitmap[8];
